local-pollution-reporting/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
supabase==2.6.0
pillow==11.0.0
requests==2.32.3
pytest==8.3.3
